requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-jazzmin==3.0.1
django-parler==2.3
djangorestframework==3.16.1
drf-flex-fields==1.0.2
idna==3.11
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
